feat: my 연관 페이지 라우터 세팅
