+++ src/App.jsx
@@ -5,9 +5,10 @@ import Like from './pages/Like'
 import Oneul from './pages/Oneul'
 import Login from './pages/Login'
 import Cart from './pages/Cart'
-import MyPage from './pages/MyPage'
-import MyReview from './pages/MyReview'
-import MyReviewWrite from './pages/MyReviewWrite'
+import My from './pages/My'
+import MyPage from './pages/My/MyPage'
+import MyReview from './pages/My/MyReview'
+import MyReviewWrite from './pages/My/MyReviewWrite'
 import SignUp from './pages/SignUp'
 import Category from './pages/Category'
 import NavBottom from './components/layout/NavBottom'
@@ -17,11 +18,12 @@ import SignUpPw from './pages/SignUp/SignUpPw'
 import UserInfo from './pages/SignUp/UserInfo'
 import Finish from './pages/SignUp/Finish'
 import Product from './pages/Product'
-import ProductDetail from './pages/ProductDetail'
+import ProductDetail from './pages/Prdoucts/ProductDetail'
 import SnapDetail from './pages/SnapDetail'
 import MagazineDetail from './pages/MagazineDetail'
 import Search from './pages/Search'
 import Order from './pages/Order'
+import QnA from './pages/My/QnA'
 
 function App() {
   const location = useLocation()
@@ -33,16 +35,18 @@ function App() {
         <Route path="/category" element={<Category />} />
         <Route path="/oneul" element={<Oneul />} />
         <Route path="/like" element={<Like />} />
-        <Route path="/my" element={<MyPage />} />
-        <Route path="/my/review" element={<MyReview />} />
-        <Route path="/my/review/write" element={<MyReviewWrite />} />
+        <Route path="/my" element={<My />}>
+          <Route path="" element={<MyPage />} />
+          <Route path="qna" element={<QnA />} />
+          <Route path="review" element={<MyReview />} />
+          <Route path="review/write" element={<MyReviewWrite />} />
+        </Route>
         <Route path="/login" element={<Login />} />
         <Route path="/product" element={<Product />} />
         {/*실제사용 아래 임시 <Route path="/product/detail/:id" element={<DetailPage />} /> */}
         <Route path="/product/detail" element={<ProductDetail />} />
         <Route path="/oneul/detail" element={<SnapDetail />} />
         <Route path="/magazine/id" element={<MagazineDetail />} />
-        <Route path="/my/review" element={<MyReview />} />
         <Route path="/signup" element={<SignUp />}>
           <Route path="" element={<TermService />} />
           <Route path="idform" element={<SignUpId />} />
