+++ src/pages/My/MyReviewWrite.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import Write from '../../components/MyReviewSegment/Write'
+const MyReviewWrite = () => {
+  return (
+    <>
+      <Write />
+    </>
+  )
+}
+
+export default MyReviewWrite
