+++ src/pages/My/QnA.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const QnA = () => {
+  return <div>QnA ififdkfmsek</div>
+}
+
+export default QnA
